Reorder ProblemReportViewController and get rid of swiftlint warnings
